Handle parenthesized $range step literals
